pyproject.toml
[tool.poetry]
name = "sms"
version = "0.1.0"
description = ""
authors = ["Razan Armouti <razanarmote@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
asgiref = "^3.4.1"
django-extensions = "^3.1.5"
graphviz = "^0.19.1"
mysql-connector-python = "^8.0.27"
protobuf = "3.12.2"
pydot = "^1.4.2"
pyparsing = "^3.0.6"
pytz = "^2021.3"
six = "^1.16.0"
sqlparse = "^0.4.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-environ = "^0.8.1"
django-cors-headers = "^3.10.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.2"
PyJWT = "^2.3.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
